requirements/base.txt
asgiref==3.7.2
async-timeout==4.0.2
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
marshmallow==3.19.0
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
transliterate==1.10.2
typing_extensions==4.7.1
tzdata==2023.3
